corrections and extensions for double datatype

+++ src/library/field.cpp
@@ -41,10 +41,12 @@ template <>
 string convert<const float&, string>(const float& value) {
     return toString(value);
 }
+
 template <>
 string convert<const double&, string>(const double& value) {
     return toString(value);
 }
+
 template <>
 string convert<const bool&, string>(const bool& value) {
     return toString(value);
@@ -62,12 +64,12 @@ template <>
 float convert<int, float>(int value) {
     return (float)value;
 }
-/*
+
 template <> 
-float convert<double, float>(double value) {
-    return (float)value;
+double convert<int,double>(int value) {
+    return (double)value;
 }
-*/
+
 template <>
 bool convert<const string&, bool>(const string& value) {
     return convert<const string&, int>(value);
